+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # BC Security Changelog
 
-## Upcoming version 0.21.0 (????-??-??)
+## Upcoming version 0.21.0 (2023-08-17)
 
-PHP 8.0 or newer and WordPress 6.0 or newer are now required. Plugin has been tested with WordPress 6.3.
+PHP 8.0 or newer and WordPress 6.0 or newer are now required!
 
 ### Added